reorganizando a pasta de teste
